+++ app/src/main/java/ru/yandex/practicum/contacts/presentation/sort/SortTypeUI.java
@@ -2,28 +2,19 @@ package ru.yandex.practicum.contacts.presentation.sort;
 
 import androidx.annotation.NonNull;
 
-public class SortTypeUI {
+import ru.yandex.practicum.contacts.presentation.base.ContactsOrderTypeUi;
 
-    private final String type;
-    private final boolean isSelected;
+public class SortTypeUI extends ContactsOrderTypeUi {
 
     public SortTypeUI(@NonNull String type, boolean isSelected) {
-        this.type = type;
-        this.isSelected = isSelected;
+        super(type,isSelected);
+        //this.type = type;
+        // this.isSelected = isSelected;
     }
-
-    public String getType() {
-        return type;
-    }
-
-    public boolean isSelected() {
-        return isSelected;
-    }
-
-    public String createLogMessage() {
+    @Override
+    public String createLogMessage(){
         return "Выбран тип сортировки: " + type;
     }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) {
